feat: add a new context to store the books data in local storage once fetched
This will reduce the number of requests send to the lambda function to 1
as long as the books data available in the localstorage
